requirements.txt
fastapi
uvicorn[standard]
langgraph
pydantic
python-multipart
requests
pinecone
sentence-transformers
supabase
pdfplumber
pymupdf
python-docx
pillow
pytesseract
aiofiles
httpx
groq
sqlalchemy
psycopg2-binary
python-dotenv
loguru